[OAP-1835][oap-native-sql] Support ColumnarBHJ to build and broadcast hashrelation (#1836)
* [oap-native-sql] Use hashmap to store all ArrayItemIndex chains instead of using an extra vector


* [oap-native-sql] Support non-codegen ProbeJoin using hash_map_type 2


* [oap-native-sql] Scala support for BHJ type2

In this type, we can broadcast the whole hashmap and use them in join side without build it


* [oap-native-sql] Refine hashmap and unsafeRow to reduce size


* [oap-native-sql] Use externalizer or kyroSerializer to serialize Broadcast obj


* [oap-native-sql] HashMap optimize for Semi, Anti and Existence


* [oap-native-sql] Skip making unsafeRow when there is single field key

When single field key dataType is smaller than uint32_t, we won't do key comparison


* [oap-native-sql] Remove breakdown logs


* [oap-native-sql] Support StringType single key in HashMap


* [oap-native-sql] Add an optimization for HashMap when key is fixedSize

when single key with Fixed Size, we will put key in keyArray to speed up


* [oap-native-sql] Add a FutureClose to close a broadcast cached obj after 30s no use


* [oap-native-sql] DataToArrow should now treat received data as ColumnarHashedRelation


* [oap-native-sql] Bug fix in expression codegen substr


* [oap-native-sql] Add new kernel ConcatArrayList used to concat multiple arrays as one


* [oap-native-sql] Support to rewriteKey optimization when multiple key total length is less than 8 bytes


* [oap-native-sql] implement KnownSizeEstimation in ColumnarHashedRelation

By doing so, memorystore could get the estimatedSize of ColumnarHashedRelation Object


* [oap-native-sql] Make Broadcast deserialize memory allocation off record


* [oap-native-sql] Fix a standalone ColumnarBHJ bug

There is a corner case bug that standalone ColumnarBHJ may use its child schema to do codegen, fix in this commit

+++ oap-native-sql/cpp/src/codegen/common/hash_relation.h
@@ -102,10 +102,17 @@ class TypedHashRelationColumn<DataType, enable_if_string_like<DataType>>
   std::vector<std::shared_ptr<StringArray>> array_vector_;
 };
 
+template <typename T>
+using is_number_alike =
+    std::integral_constant<bool, std::is_arithmetic<T>::value ||
+                                     std::is_floating_point<T>::value>;
+
 /////////////////////////////////////////////////////////////////////////
 
 class HashRelation {
  public:
+  HashRelation(arrow::compute::FunctionContext* ctx) {}
+
   HashRelation(
       const std::vector<std::shared_ptr<HashRelationColumn>>& hash_relation_list) {
     hash_relation_column_list_ = hash_relation_list;
@@ -113,13 +120,14 @@ class HashRelation {
 
   HashRelation(
       arrow::compute::FunctionContext* ctx,
-      const std::vector<std::shared_ptr<HashRelationColumn>>& hash_relation_column)
+      const std::vector<std::shared_ptr<HashRelationColumn>>& hash_relation_column,
+      int key_size = -1)
       : HashRelation(hash_relation_column) {
-    hash_table_ = createUnsafeHashMap(1024 * 1024 * 4, 1024 * 1024);
+    hash_table_ = createUnsafeHashMap(1024 * 1024, 256 * 1024 * 1024, key_size);
   }
 
   ~HashRelation() {
-    if (hash_table_ != nullptr) {
+    if (hash_table_ != nullptr && !unsafe_set) {
       destroyHashMap(hash_table_);
       hash_table_ = nullptr;
     }
@@ -134,48 +142,116 @@ class HashRelation {
       const std::vector<std::shared_ptr<UnsafeArray>>& payloads) {
     // This Key should be Hash Key
     auto typed_array = std::make_shared<ArrayType>(in);
-    if (typed_array->null_count() == 0) {
-      for (int i = 0; i < typed_array->length(); i++) {
-        std::shared_ptr<UnsafeRow> payload = std::make_shared<UnsafeRow>();
-        initTempUnsafeRow(payload.get(), payloads.size());
-        for (auto payload_arr : payloads) {
-          payload_arr->Append(i, &payload);
-        }
-        RETURN_NOT_OK(Insert(typed_array->GetView(i), payload, num_arrays_, i));
-        releaseTempUnsafeRow(payload.get());
+    for (int i = 0; i < typed_array->length(); i++) {
+      std::shared_ptr<UnsafeRow> payload = std::make_shared<UnsafeRow>(payloads.size());
+      for (auto payload_arr : payloads) {
+        payload_arr->Append(i, &payload);
       }
-    } else {
-      for (int i = 0; i < typed_array->length(); i++) {
-        std::shared_ptr<UnsafeRow> payload = std::make_shared<UnsafeRow>();
-        initTempUnsafeRow(payload.get(), payloads.size());
-        for (auto payload_arr : payloads) {
-          payload_arr->Append(i, &payload);
-        }
-        if (typed_array->IsNull(i)) {
-          RETURN_NOT_OK(InsertNull(num_arrays_, i));
-        } else {
-          RETURN_NOT_OK(Insert(typed_array->GetView(i), payload, num_arrays_, i));
-        }
-        releaseTempUnsafeRow(payload.get());
-      }
+      RETURN_NOT_OK(Insert(typed_array->GetView(i), payload, num_arrays_, i));
     }
+
     num_arrays_++;
+    // DumpHashMap();
     return arrow::Status::OK();
   }
 
-  int Get(int32_t v, std::shared_ptr<UnsafeRow> payload) {
-    assert(hash_table_ != nullptr);
-    return safeLookup(hash_table_, payload.get(), v);
+  template <typename KeyArrayType>
+  arrow::Status AppendKeyColumn(std::shared_ptr<arrow::Array> in,
+                                std::shared_ptr<KeyArrayType> original_key) {
+    // This Key should be Hash Key
+    auto typed_array = std::make_shared<ArrayType>(in);
+    for (int i = 0; i < typed_array->length(); i++) {
+      RETURN_NOT_OK(
+          Insert(typed_array->GetView(i), original_key->GetView(i), num_arrays_, i));
+    }
+
+    num_arrays_++;
+    // DumpHashMap();
+    return arrow::Status::OK();
   }
 
-  int GetNull() {
-    if (!null_index_set_) {
-      return HASH_NEW_KEY;
-    } else {
-      auto ret = null_index_;
-      return ret;
+  arrow::Status AppendKeyColumn(std::shared_ptr<arrow::Array> in,
+                                std::shared_ptr<StringArray> original_key) {
+    // This Key should be Hash Key
+    auto typed_array = std::make_shared<ArrayType>(in);
+    for (int i = 0; i < typed_array->length(); i++) {
+      RETURN_NOT_OK(
+          Insert(typed_array->GetView(i), original_key->GetString(i), num_arrays_, i));
     }
+
+    num_arrays_++;
+    // DumpHashMap();
+    return arrow::Status::OK();
   }
+
+  template <typename CType,
+            typename std::enable_if_t<is_number_alike<CType>::value>* = nullptr>
+  int Get(int32_t v, CType payload) {
+    if (hash_table_ == nullptr) {
+      throw std::runtime_error("HashRelation Get failed, hash_table is null.");
+    }
+    std::vector<char*> res_out;
+    auto res = safeLookup(hash_table_, payload, v, &res_out);
+    if (res == -1) return -1;
+    arrayid_list_.clear();
+    for (auto index : res_out) {
+      arrayid_list_.push_back(*((ArrayItemIndex*)index));
+    }
+    return 0;
+  }
+
+  int Get(int32_t v, std::string payload) {
+    if (hash_table_ == nullptr) {
+      throw std::runtime_error("HashRelation Get failed, hash_table is null.");
+    }
+    std::vector<char*> res_out;
+    auto res = safeLookup(hash_table_, payload.data(), payload.size(), v, &res_out);
+    if (res == -1) return -1;
+    arrayid_list_.clear();
+    for (auto index : res_out) {
+      arrayid_list_.push_back(*((ArrayItemIndex*)index));
+    }
+    return 0;
+  }
+
+  int Get(int32_t v, std::shared_ptr<UnsafeRow> payload) {
+    if (hash_table_ == nullptr) {
+      throw std::runtime_error("HashRelation Get failed, hash_table is null.");
+    }
+    std::vector<char*> res_out;
+    auto res = safeLookup(hash_table_, payload, v, &res_out);
+    if (res == -1) return -1;
+    arrayid_list_.clear();
+    for (auto index : res_out) {
+      arrayid_list_.push_back(*((ArrayItemIndex*)index));
+    }
+    return 0;
+  }
+
+  template <typename CType,
+            typename std::enable_if_t<is_number_alike<CType>::value>* = nullptr>
+  int IfExists(int32_t v, CType payload) {
+    if (hash_table_ == nullptr) {
+      throw std::runtime_error("HashRelation Get failed, hash_table is null.");
+    }
+    return safeLookup(hash_table_, payload, v);
+  }
+
+  int IfExists(int32_t v, std::string payload) {
+    if (hash_table_ == nullptr) {
+      throw std::runtime_error("HashRelation Get failed, hash_table is null.");
+    }
+    return safeLookup(hash_table_, payload.data(), payload.size(), v);
+  }
+
+  int IfExists(int32_t v, std::shared_ptr<UnsafeRow> payload) {
+    if (hash_table_ == nullptr) {
+      throw std::runtime_error("HashRelation Get failed, hash_table is null.");
+    }
+    return safeLookup(hash_table_, payload, v);
+  }
+
+  int GetNull() { return null_index_set_ ? 0 : HASH_NEW_KEY; }
 
   arrow::Status AppendPayloadColumn(int idx, std::shared_ptr<arrow::Array> in) {
     return hash_relation_column_list_[idx]->AppendColumn(in);
@@ -191,26 +267,68 @@ class HashRelation {
     return arrow::Status::OK();
   }
 
-  std::vector<ArrayItemIndex> GetItemListByIndex(int i) {
-    return memo_index_to_arrayid_[i];
+  arrow::Status UnsafeGetHashTableObject(int64_t* addrs, int* sizes) {
+    if (hash_table_ == nullptr) {
+      return arrow::Status::Invalid("UnsafeGetHashTableObject hash_table is null");
+    }
+    // dump(hash_table_);
+    addrs[0] = (int64_t)hash_table_;
+    sizes[0] = (int)sizeof(unsafeHashMap);
+
+    addrs[1] = (int64_t)(hash_table_->keyArray);
+    sizes[1] = (int)(hash_table_->arrayCapacity * hash_table_->bytesInKeyArray);
+
+    addrs[2] = (int64_t)(hash_table_->bytesMap);
+    sizes[2] = (int)(hash_table_->cursor);
+    return arrow::Status::OK();
   }
 
-  uint64_t GetHashRelationLength() { return num_items_; }
+  arrow::Status UnsafeSetHashTableObject(int len, int64_t* addrs, int* sizes) {
+    assert(len == 3);
+    hash_table_ = (unsafeHashMap*)addrs[0];
+    hash_table_->cursor = sizes[2];
+    hash_table_->keyArray = (char*)addrs[1];
+    hash_table_->bytesMap = (char*)addrs[2];
+    unsafe_set = true;
+    // dump(hash_table_);
+    return arrow::Status::OK();
+  }
+
+  arrow::Status DumpHashMap() {
+    dump(hash_table_);
+    return arrow::Status::OK();
+  }
+
+  virtual std::vector<ArrayItemIndex> GetItemListByIndex(int i) { return arrayid_list_; }
+
+  void TESTGrowAndRehashKeyArray() { growAndRehashKeyArray(hash_table_); }
 
  protected:
-  uint64_t num_items_ = 0;
+  bool unsafe_set = false;
   uint64_t num_arrays_ = 0;
-  std::vector<std::vector<ArrayItemIndex>> memo_index_to_arrayid_;
   std::vector<std::shared_ptr<HashRelationColumn>> hash_relation_column_list_;
+  unsafeHashMap* hash_table_ = nullptr;
+  using ArrayType = sparkcolumnarplugin::precompile::Int32Array;
+  bool null_index_set_ = false;
+  std::vector<ArrayItemIndex> null_index_list_;
+  std::vector<ArrayItemIndex> arrayid_list_;
+
   arrow::Status Insert(int32_t v, std::shared_ptr<UnsafeRow> payload, uint32_t array_id,
                        uint32_t id) {
     assert(hash_table_ != nullptr);
-    int i = getOrInsert(hash_table_, payload.get(), v, num_items_);
-    if (i < num_items_) {
-      memo_index_to_arrayid_[i].emplace_back(array_id, id);
-    } else {
-      num_items_++;
-      memo_index_to_arrayid_.push_back({ArrayItemIndex(array_id, id)});
+    auto index = ArrayItemIndex(array_id, id);
+    if (!append(hash_table_, payload.get(), v, (char*)&index, sizeof(ArrayItemIndex))) {
+      return arrow::Status::CapacityError("Insert to HashMap failed.");
+    }
+    return arrow::Status::OK();
+  }
+
+  template <typename CType>
+  arrow::Status Insert(int32_t v, CType payload, uint32_t array_id, uint32_t id) {
+    assert(hash_table_ != nullptr);
+    auto index = ArrayItemIndex(array_id, id);
+    if (!append(hash_table_, payload, v, (char*)&index, sizeof(ArrayItemIndex))) {
+      return arrow::Status::CapacityError("Insert to HashMap failed.");
     }
     return arrow::Status::OK();
   }
@@ -218,18 +336,12 @@ class HashRelation {
   arrow::Status InsertNull(uint32_t array_id, uint32_t id) {
     if (!null_index_set_) {
       null_index_set_ = true;
-      null_index_ = num_items_++;
-      memo_index_to_arrayid_.push_back({ArrayItemIndex(array_id, id)});
+      null_index_list_ = {ArrayItemIndex(array_id, id)};
     } else {
-      memo_index_to_arrayid_[null_index_].emplace_back(array_id, id);
+      null_index_list_.emplace_back(array_id, id);
     }
     return arrow::Status::OK();
   }
-
-  unsafeHashMap* hash_table_ = nullptr;
-  using ArrayType = sparkcolumnarplugin::precompile::Int32Array;
-  bool null_index_set_ = false;
-  int32_t null_index_;
 };
 
 template <typename T, typename Enable = void>
